backend/requirements.txt
livekit-agents
livekit-plugins-openai>=0.12.0
livekit-plugins-noise-cancellation
requests
python-dotenv
pydantic-ai-slim[openai,mcp]
livekit-agents[tavus]~=1.0
tzdata
flask[async]
flask
flask-cors